Bird is indeed the wordddddd
